Fix load flow model extraction
